stack: improve tests on 32 bits
Reenable examples, changed float64 to float32 and it works just fine.
Skip the float64 test on 32 bits until parsing is improved.

Helps with issue #65.

+++ stack/example_test.go
@@ -5,9 +5,6 @@
 // Output is currently different on 32 bits so skip the examples for now. It's
 // a gross hack.
 
-//go:build !386 && !arm && !mips && !mipsle
-// +build !386,!arm,!mips,!mipsle
-
 package stack_test
 
 import (
@@ -41,7 +38,7 @@ func Example_text() {
 
 	type crashy struct{}
 
-	func (c crashy) die(f float64) {
+	func (c crashy) die(f float32) {
 		time.Sleep(time.Millisecond)
 		panic(int(f))
 	}`
